refactor: change the formation of paths for API functions
