+++ async-race/src/constants/routes.ts
@@ -1,15 +1,3 @@
-import { Category, OptionsStatus, Path } from '../types/type';
-
-export const URL_API = 'http://127.0.0.1:3000';
-
-export const PATH_MAP: Record<Category, Path> = {
-  garage: '/garage',
-  engine: '/engine',
-  winners: '/winners',
-};
-
-export const OPTIONS_MAP: Record<OptionsStatus, OptionsStatus> = {
-  started: 'started',
-  drive: 'drive',
-  stopped: 'stopped',
-};
+export const URL_GARAGE = 'http://127.0.0.1:3000/garage/';
+export const URL_ENGINE = 'http://127.0.0.1:3000/engine/';
+export const URL_WINNERS = 'http://127.0.0.1:3000/winners/';
