+++ async-race/src/types/type.ts
@@ -1,5 +1,3 @@
-export type Path = '/garage' | '/engine' | '/winners';
-export type Category = 'garage' | 'engine' | 'winners';
 export type OptionsStatus = 'started' | 'stopped' | 'drive';
 export type Action = 'add' | 'remove' | 'add-100';
