+++ async-race/src/view/winnersPage.ts
@@ -1,8 +1,9 @@
 import { getCarAPI, getWinnersAPI } from '../api/api';
 import { Winners } from '../types/type';
-import { NAME_HEADER_TABLE, START_PAGE_GARAGE, WINNERS_ON_PAGE } from '../constants/common';
+import { NAME_HEADER_TABLE } from '../constants/common';
 import { createButton, createElement } from '../utils/utils';
 import carImg from '../assets/car.gif';
+import { START_PAGE_GARAGE, WINNERS_ON_PAGE } from '../constants/api';
 
 export async function createWinCar(elem: Winners): Promise<void> {
   const ID = elem.id;
